Refactored v3 heuristic gas model
